+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Update `User-Agent` and resources
 * Add new features and improve prediction quality
 * Remove de-duplication of battles (not needed anymore), filter out battles without hero powers
+* Change logging format, remove emoji's
 
 # `2.0`